Initial clean up and new file
